Refactor: ダッシュボード内のコンポーネントを `DashboardItems` に分離
